Update PlayerCommand.js
#121

+++ src/modules/PlayerCommand.js
@@ -1,4 +1,3 @@
-var Entity = require('../entity');
 var Logger = require('./Logger');
 var UserRoleEnum = require("../enum/UserRoleEnum");
 
@@ -67,11 +66,19 @@ PlayerCommand.prototype.userLogin = function (ip, password) {
 
 var playerCommands = {
     help: function (args) {
+        if (this.playerTracker.userRole != (UserRoleEnum.ADMIN || UserRoleEnum.MODER || UserRoleEnum.USER)) {
+            this.writeLine("ERROR: access denied!");
+            return;
+        }
         this.writeLine("/skin %shark - change skin");
         this.writeLine("/kill - self kill");
         this.writeLine("/help - this command list");
     },
     skin: function (args) {
+        if (this.playerTracker.userRole != (UserRoleEnum.ADMIN || UserRoleEnum.MODER || UserRoleEnum.USER)) {
+            this.writeLine("ERROR: access denied!");
+            return;
+        }
         if (this.playerTracker.cells.length > 0) {
             this.writeLine("ERROR: Cannot change skin while player in game!");
             return;
@@ -85,6 +92,10 @@ var playerCommands = {
             this.writeLine("Your skin set to " + skinName);
     },
     kill: function (args) {
+        if (this.playerTracker.userRole != (UserRoleEnum.ADMIN || UserRoleEnum.MODER || UserRoleEnum.USER)) {
+            this.writeLine("ERROR: access denied!");
+            return;
+        }
         if (this.playerTracker.cells.length < 1) {
             this.writeLine("You cannot kill yourself, because you're still not joined to the game!");
             return;
